BLE: Move BleInstanceBase.h into source.
