+++ connectivity/FEATURE_BLE/libraries/TARGET_CORDIO/driver/CordioHCIDriver.cpp
@@ -18,7 +18,6 @@
 #include <string.h>
 
 #include "ble/BLE.h"
-#include "ble/internal/BLEInstanceBase.h"
 #include "ble/Gap.h"
 #include "CordioHCIDriver.h"
 #include "hci_api.h"
@@ -98,7 +97,7 @@ buf_pool_desc_t CordioHCIDriver::get_default_buffer_pool_description()
 
 void CordioHCIDriver::set_random_static_address(const ble::address_t& address)
 {
-    ble_error_t err = BLEInstanceBase::deviceInstance().getGap().setRandomStaticAddress(address);
+    ble_error_t err = BLE::Instance().gap().setRandomStaticAddress(address);
     MBED_ASSERT(err == BLE_ERROR_NONE);
 }
