+++ connectivity/FEATURE_BLE/libraries/TARGET_CORDIO/include/ble/internal/BLEInstanceBaseImpl.h
@@ -21,7 +21,7 @@
 
 #include "ble/BLE.h"
 #include "ble/common/ble/blecommon.h"
-#include "ble/internal/BLEInstanceBase.h"
+#include "source/BLEInstanceBase.h"
 
 #include "CordioHCIDriver.h"
 #include "ble/GattServer.h"
