+++ connectivity/drivers/ble/FEATURE_BLE/COMPONENT_BlueNRG_MS/BlueNrgMsHCIDriver.cpp
@@ -27,7 +27,6 @@
 #include "platform/mbed_wait_api.h"
 
 // FEATURE_BLE/targets/TARGET_CORDIO
-#include "BLEInstanceBase.h"
 #include "CordioHCIDriver.h"
 #include "CordioHCITransportDriver.h"
 #include "hci_api.h"
