+++ connectivity/drivers/ble/FEATURE_BLE/COMPONENT_CYW43XXX/HCIDriver.cpp
@@ -16,7 +16,6 @@
  */
 
 #include <stdio.h>
-#include "BLEInstanceBase.h"
 #include "CordioHCIDriver.h"
 #include "hci_api.h"
 #include "hci_cmd.h"
