+++ connectivity/drivers/ble/FEATURE_BLE/COMPONENT_CYW43XXX/TARGET_PSOC6/cy_bt_cordio_cfg.cpp
@@ -24,7 +24,6 @@
 *******************************************************************************/
 
 #include <stdio.h>
-#include "BLEInstanceBase.h"
 #include "CordioHCIDriver.h"
 #include "hci_api.h"
 #include "hci_cmd.h"
